Removed global project name
I realized that having all different services grouped together makes
things worse, not better. They are suddenly all in the default network
together with all other services, unless configured otherwise. I always
get the message that there are orphaned containers, when starting a part
of the project. And I realized, that I won't really be able to run
multiple projects like this one together on the same server.And there
aren't just any good benefits to justify these inconveniences.

So with this, I revert back to be treating the individual parts of this
project as that: Individual. They might share config for traefik, or
watchtower, when they need to, but otherwise things are separated.

I kept the manage.sh as it makes it very convenient to work with this
setup.