Migrated from EditorConfig to Prettier
I am currently using Zed, which does not support editorconf. Migrating
to Prettier allows me to control the auto-formatting behaviour of Zed
in the same way as I could with .editorconf in VS-Code.

This is important as in the next commits there will a complete
refactoring of this project. I want the formatting to stay the same so
the git log is more readable.

For some reason, the single-quote setting did not take effect when using
EditorConfig, so I needed to add an override for yaml files to preserve
the correct quotation. I might change that back in another commit that
does not incoorporate any other changes.