Next Verion CloudBaby
Here I heavily refactored the setup for my cloud server.

Most notably I decided to stick with docker compose, rather than
migrating to stack. Stack lacks hardware accelleration support and was
more cumbersome to use for me. Restoring a database backup for example
was not working as expected and I could not get it to work. Also the
promise of remote deployments is not as relevant for me when using the
zeditor to directly work on the remote machine. I also prefer sshing
into the machine. Wich docker stack and context, I multiple times
deployed services to the wrong machine.

I also threw out the weird manage.sh script that I always forgot how to
use. I feel it wasn't that mature and intuitive to me. In it's place are now
Makefiles, which also serve as documentation. I can set env vars there
and even compute the trusted proxy ip for NextCloud in them. So all
configuration is in one place, which I like

I also removed most of the env vars. Some, I replaces with docker
compose secrets, as I can't have them in the Makefiles, because those
are checked in obviously.
The other env vars are now just inlined into the docker-compose.yml
files. This makes adapting them much more easier. I now only use env
vars if a value is used in multiple places in the docker-compose.yml
or if there is a similar value set as an env var.
This makes adapting the setup more enjoyable. I can do most things in
the docker-compose.yml files and don't need to search and jump around to
find values.
I realized I don't need the additional flexibility of all the env vars.
In the end I can always adapt the docker-compose.yml files. I probably
will never run the same setup on a different server so that I need to
change the env vars, and even if, I'd probably clone the repo anyway as
I would need a different set of services in such a case.