Consistent yaml style
I never had a consistent style in the yaml files here. The reason for
this was, that the files were copy and pasted from the internet and me
having no strict formatter.

I still don't have that formatter.

But I now have a consistent style.

I opted for explicitly defining the type of data everywhere. Strings are
always quoted, even most don't need to be. I like this as it is a simple
rule to follow when working with these files.

I adapted all yaml files in the project.